# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Haife API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/public/login": {
      "post": {
        "tags": [
          "Public"
        ],
        "summary": "系統登入",
        "description": "傳入帳號密碼並進行驗證登入",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "account": {
                  "type": "string",
                  "example": "帳號"
                },
                "password": {
                  "type": "string",
                  "example": "密碼"
                }
              },
              "required": [
                "account",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登入成功"
          },
          "400": {
            "description": "登入失敗"
          },
          "403": {
            "description": "無法通過帳密驗證"
          }
        }
      }
    },
...
    "/api/prizes": {
      "post": {
        "tags": [
          "Prizes"
        ],
        "summary": "新增獎項",
        "description": "傳入獎項資訊並建立獎項",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "獎項新增資訊",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "獎項名稱"
                },
                "price": {
                  "type": "string",
                  "example": "價格"
                },
                "time": {
                  "type": "string",
                  "example": "個數"
                },
                "quantity": {
                  "type": "string",
                  "example": "獎項數量"
                },
                "remainingQty": {
                  "type": "string",
                  "example": "剩餘數量"
                },
                "type": {
                  "type": "string",
                  "example": "獎項類型"
                },
                "isEnabled": {
                  "type": "string",
                  "example": "是否啟用"
                },
                "activity_uuid": {
                  "type": "string",
                  "example": "活動編號"
                }
              },
              "required": [
                "name",
                "price",
                "time",
                "quantity",
                "remainingQty",
                "type",
                "isEnabled",
                "activity_uuid"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "新增成功"
          },
          "400": {
            "description": "新增失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
    "/api/prizes/{uuid}": {
      "patch": {
        "tags": [
          "Prizes"
        ],
        "summary": "修改獎項",
        "description": "根據獎項編號，傳入獎項資訊並修改獎項",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "獎項編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "獎項修改資訊",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "獎項名稱"
                },
                "price": {
                  "type": "string",
                  "example": "價格"
                },
                "time": {
                  "type": "string",
                  "example": "個數"
                },
                "quantity": {
                  "type": "string",
                  "example": "獎項數量"
                },
                "remainingQty": {
                  "type": "string",
                  "example": "剩餘數量"
                },
                "type": {
                  "type": "string",
                  "example": "獎項類型"
                },
                "isEnabled": {
                  "type": "string",
                  "example": "是否啟用"
                }
              },
              "required": [
                "name",
                "price",
                "time",
                "quantity",
                "remainingQty",
                "type",
                "isEnabled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "400": {
            "description": "修改失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      },
      "delete": {
        "tags": [
          "Prizes"
        ],
        "summary": "刪除獎項",
        "description": "根據獎項編號刪除獎項",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "獎項編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "刪除成功"
          },
          "400": {
            "description": "刪除失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
    "/api/prizes/{activity_uuid}": {
      "get": {
        "tags": [
          "Prizes"
        ],
        "summary": "獎項查詢 (含分頁與搜尋)",
        "description": "根據活動編號列出所有獎項資訊 (含分頁與搜尋)",
        "parameters": [
          {
            "name": "activity_uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "活動編號"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "rowsPerPage",
            "in": "query",
            "required": false,
            "description": "每頁筆數",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "目前頁數",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "排序欄位",
            "type": "string"
          },
          {
            "name": "sortType",
            "in": "query",
            "required": false,
            "description": "排序方式",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "搜尋資訊 (可空值)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
...
    "/api/draw/prizes": {
      "get": {
        "tags": [
          "Draw"
        ],
        "summary": "取得所有獎項",
        "description": "列出所有獎項",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "抽獎序號",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          }
        }
      }
    },
...
  }
}